rename and refactor
renamed first-seldon-deployment to  first-seldon-core-microservice to first-seldon and moved kubernetes related information to a new post titled deploy-first-microservice.
This was done to start modularizing the posts. Also, the user has already accomplished something at building the dockerfile, so I want it to feel that way.

Kubernetes will cause more difficulty, so it's nice to add a break in there